Upgrade Lambda function to Python Cloudflare v4
