+++ .github/workflows/dev.yml
@@ -13,7 +13,8 @@ env:
   TF_VAR_org_primary_account: ${{ secrets.ORG_PRIMARY_ACCOUNT }}
   TF_VAR_slack_webhook_urls: ${{ secrets.SLACK_WEBHOOK_URLS }}
   TF_VAR_external_id: ${{ secrets.EXTERNAL_ID }}
-  TF_VAR_cf_api_key: ${{ secrets.CF_API_KEY }}
+  TF_VAR_cloudflare_api_key: ${{ secrets.CLOUDFLARE_API_KEY }}
+  TF_VAR_cloudflare_email: ${{ secrets.CLOUDFLARE_EMAIL }}
   TF_VAR_hackerone_api_token: ${{ secrets.HACKERONE_API_TOKEN }}
   TF_VAR_region: ${{ secrets.AWS_REGION }}
   TF_CLI_ARGS_init: "-backend-config=\"bucket=${{ secrets.TERRAFORM_STATE_BUCKET }}\" -backend-config=\"key=${{ secrets.TERRAFORM_STATE_KEY }}\" -backend-config=\"region=${{ secrets.TERRAFORM_STATE_REGION }}\""
