+++ .github/workflows/integration_tests_deployment.yml
@@ -50,7 +50,7 @@ jobs:
 
       - name: Install dependencies
         run: |
-          pip install -r requirements-tests.txt
+          pip install -r requirements-dev.txt
 
       - name: Configure AWS Credentials
         uses: aws-actions/configure-aws-credentials@e3dd6a429d7300a6a4c196c26e071d42e0343502 # v4
