+++ requirements-dev.txt
@@ -1,5 +1,5 @@
 boto3==1.34.25
-cloudflare==2.20.0
+cloudflare==4.0.0
 dnspython==2.7.0
 requests==2.32.3
 black==24.10.0
