--- requirements-tests.txt
@@ -1,13 +1,0 @@
-boto3==1.34.25
-cloudflare==4.0.0
-dnspython==2.7.0
-requests==2.32.3
-black==24.10.0
-prospector==1.13.3
-pytest==8.3.4
-pytest-cov==6.0.0
-assertpy==1.1
-moto[sts, iam]==4.2.14
-requests-mock==1.12.1
-checkov==3.2.286
-pre-commit==4.0.1
