+++ utils/utils_cloudflare.py
@@ -1,59 +1,42 @@
-import CloudFlare
+# Cloudflare Python version 4.0.0 - DNS Utils
+from cloudflare import Cloudflare
+from cloudflare import CloudflareError
 
 
-def list_cloudflare_zones():
+def list_dns_records(zone_id, zone_name, record_fqdn=None):
+    cf = Cloudflare()
+    print(f"Listing DNS records in Cloudflare DNS zone {zone_name}")
 
-    cf = CloudFlare.CloudFlare(raw=True)
-    print("Searching for DNS zones in Cloudflare")
-
-    page_number = 0
-    total_zones = 0
-    zone_list = []
-
-    while True:
-        page_number += 1
-        raw_results = cf.zones.get(params={"per_page": 5, "page": page_number})
-        zones = raw_results["result"]
-
-        for zone in zones:
-            total_zones += 1
-            zone_id = zone["id"]
-            zone_name = zone["name"]
-            zone_list.append({"Id": zone_id, "Name": zone_name})
-
-        zone_total_pages = raw_results["result_info"]["total_pages"]
-        if page_number == zone_total_pages:
-            break
-
-    print(f"returned {total_zones} Cloudflare DNS zones")
-
-    return zone_list
+    try:
+        records = cf.dns.records.list(zone_id=zone_id, name=record_fqdn).result
+        print(f"Successfully listed DNS records in Cloudflare DNS zone {zone_name}")
+        return records
+    except CloudflareError as e:
+        print(f"Failed to list DNS records in Cloudflare DNS zone {zone_name}")
+        print(f"Error: {e}")
+        return None
 
 
-def list_cloudflare_records(zone_id, zone_name):
+def list_dns_zones():
+    cf = Cloudflare()
+    print("Listing DNS zones in Cloudflare")
 
-    cf = CloudFlare.CloudFlare(raw=True)
-    print(f"Searching for DNS records in Cloudflare DNS zone {zone_name}")
+    try:
+        zones = cf.zones.list().result
+        print("Successfully listed DNS zones in Cloudflare")
+        return zones
+    except CloudflareError as e:
+        print("Failed to list DNS zones in Cloudflare")
+        print(f"Error: {e}")
+        return None
 
-    page_number = 0
-    total_records = 0
-    record_list = []
 
-    while True:
-        page_number += 1
-        raw_results = cf.zones.dns_records.get(zone_id, params={"per_page": 5, "page": page_number})
-        records = raw_results["result"]
+def convert_cf_records_to_dict(cf_records):
+    records = []
 
-        for record in records:
-            total_records += 1
-            record_list.append(
-                {"Name": record["name"], "Type": record["type"], "Value": record["content"], "Id": record["id"]},
-            )
+    for cf_record in cf_records:
+        records.append(
+            {"Name": cf_record.name, "Type": cf_record.type, "Value": cf_record.content},
+        )
 
-        zone_total_pages = raw_results["result_info"]["total_pages"]
-        if page_number == zone_total_pages:
-            break
-
-    print(f"returned {total_records} Cloudflare DNS records")
-
-    return record_list
+    return records
